# docs/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Kittisap Public + Customer API
  version: 1.0.0
  description: |
    API contract for customer-facing storefront and customer account flows.
    Admin APIs are excluded here.
servers:
  - url: https://kittisap.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasicOkResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
  schemas:
    BasicOkResponse:
      type: object
      properties:
        ok: { type: boolean, const: true }
    ErrorResponse:
      type: object
      required: [ok, code, error]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
          examples:
            - AUTH_REQUIRED
            - INVALID_REQUEST
            - OUT_OF_STOCK
            - COUPON_INVALID
        error:
          type: string
...